Move admin Cypress tests to main browser_test service
It's possible to run both these groups of tests in the same
process, which will save CI time, since we only have to do setup
and teardown once.
