+++ browser_test/cypress.catalog.json
@@ -1,0 +1,6 @@
+{
+  "baseUrl": "http://localhost:3000",
+  "video": false,
+  "viewportHeight": 2000,
+  "specPattern": "cypress/integration/catalog/**/*.spec.{js,ts}"
+}
