--- browser_test/cypress.json
@@ -1,5 +1,0 @@
-{
-  "baseUrl": "http://localhost:3000",
-  "video": false,
-  "viewportHeight": 2000
-}
