+++ browser_test/cypress/integration/catalog/contactpage.spec.js
@@ -1,4 +1,4 @@
-import terminalLog from '../utils/terminalLog'
+import terminalLog from '../../utils/terminalLog'
 
 describe("Contact", function(){
   beforeEach(() => {
