+++ browser_test/cypress/integration/catalog/homepage.spec.js
@@ -1,4 +1,4 @@
-import terminalLog from '../utils/terminalLog'
+import terminalLog from '../../utils/terminalLog'
 
 describe("HomePage", function(){
   beforeEach(() => {
